Factor out the auth middleware client
